+++ src/main/java/com/habitpay/habitpay/global/response/ApiResponse.java
@@ -1,4 +1,4 @@
-package com.habitpay.habitpay.global.common.response;
+package com.habitpay.habitpay.global.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
